update curser logic
